Use default imports for resizable and grid-layout
Changes the `@JSImport`'s to use `@JSImport.Default` and then call the import needed from the object.
Importing named imports doesn't work great for commonJS libraries and causes issues with Vite, because the import is transformed to `import * as $module from 'module'` and then the named import is accessed as `$module.myImport`.